DB設計

+++ database/migrations/2020_05_26_201735_create_tweets_table.php
@@ -14,8 +14,8 @@ class CreateTweetsTable extends Migration
     public function up()
     {
         Schema::create('tweets', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id')->comment('ユーザID');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id')->comment('ユーザID');
             $table->string('text')->comment('本文');
             $table->softDeletes();
             $table->timestamps();
